requirements.txt
langchain
openai
faiss-cpu
streamlit
tiktoken
PyPDF2
python-dotenv
